study cafe 사용량 count 함수로 변경

+++ src/main/java/com/seatly/seatly/dto/studycafe/StudyCafeUsage.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @AllArgsConstructor
 public class StudyCafeUsage {
 
-  private int totalCount;
-  private int useCount;
+  private long totalCount;
+  private long useCount;
 
 }
